tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.arc.twopie.6.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.path.arc.twopie.6</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.arc.twopie.6</h1>
<p class="desc">arc() draws correctly when start = 5 and end = start + 2pi+e and clockwise</p>


<script>
var t = async_test("arc() draws correctly when start = 5 and end = start + 2pi+e and clockwise");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#fff';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#000';
  ctx.beginPath();
  ctx.moveTo(50, 25);
  ctx.arc(50, 25, 50, 5, 5 + 2*Math.PI, false);
  ctx.closePath();
  ctx.fill();
  _assertPixel(canvas, 5,25, 0,0,0,255);
  t.done();

});
</script>
